Update catd mean to float64
To aid in merging #22, where we've collapsed the xyz dimensions into an S dimension. Tracing back on a numpy core bug with large arrays: https://github.com/numpy/numpy/issues/8869#issuecomment-369620242

+++ tedana/interfaces/tedana.py
@@ -407,7 +407,7 @@ def fitmodels_direct(catd, mmix, mask, t2s, t2sG, tes, combmode, head,
     nx, ny, nz, _, n_components = betas.shape
     n_voxels = mask.sum()
     n_data_voxels = (t2s != 0).sum()
-    mu = catd.mean(axis=-1)
+    mu = catd.mean(axis=-1, dtype=float)
     tes = np.reshape(tes, (n_echoes, 1))
     fmin, fmid, fmax = getfbounds(n_echoes)
